Add shared TabContainer
